We are getting much closer to a functional implimentation.
I had a bit of a panic where I thought torchscript was not in fact properly scripting under some circumstances, but it turned out to be a false alarm.

Other than that, throwahead is functioning and I am working towards getting attention mechanisms working

I have decided to go with the factory
class design due to all the stateful
crap that keeps popping up.